appengine/requirements.txt
flask==1.1.2
google-cloud-datastore==2.1.3
google-cloud-pubsub
